Rename Type from prisma to _Model
